requirements.txt
# TourismIQ Platform - Unified Requirements
# Python 3.11+
# Tourism Intelligence Platform with ML/AI capabilities

# ============================================
# Core Data Engineering & Processing
# ============================================
pandas==2.2.0
polars==0.20.0
duckdb==0.10.0
pyarrow==15.0.0
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2

# ============================================
# Machine Learning - Core Models
# ============================================
scikit-learn==1.4.0
lightgbm==4.3.0
xgboost==2.0.3
optuna==3.5.0

# ============================================
# NLP & Embeddings
# ============================================
transformers==4.37.0
sentence-transformers==2.3.1
langdetect==1.0.9
spacy==3.7.2
textstat==0.7.3
faiss-cpu==1.7.4

# ============================================
# ML Clustering & Dimensionality Reduction
# ============================================
hdbscan==0.8.33
umap-learn==0.5.5

# ============================================
# API & Web Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx>=0.27.0

# ============================================
# Database & Cache
# ============================================
sqlalchemy==2.0.25
redis==5.0.1
supabase==2.3.0
asyncpg==0.29.0

# ============================================
# Geospatial Processing
# ============================================
geopy==2.4.1
shapely==2.0.2

# ============================================
# Visualization & Dashboard
# ============================================
streamlit==1.50.0
plotly==6.3.1
folium==0.15.1
seaborn==0.13.1
matplotlib==3.8.2

# ============================================
# Configuration & Utilities
# ============================================
PyYAML==6.0.1
schedule==1.2.0

# ============================================
# Optional: LLM Integration (if needed)
# ============================================
# google-generativeai==0.3.2

# ============================================
# Development & Testing
# ============================================
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3
black==24.1.0
flake8==7.0.0
mypy==1.8.0
